CAN_Test_py/requirements.txt
action-msgs==1.6.0
action-tutorials-interfaces==0.27.2
action-tutorials-py==0.27.2
actionlib-msgs==5.0.1
actuator-msgs==0.0.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
ament-cmake-test==2.0.6
ament-copyright==0.14.4
ament-cppcheck==0.14.4
ament-cpplint==0.14.4
ament-flake8==0.14.4
ament-index-python==1.5.2
ament-lint==0.14.4
ament-lint-cmake==0.14.4
ament-package==0.15.3
ament-pep257==0.14.4
ament-uncrustify==0.14.4
ament-xmllint==0.14.4
angles==1.16.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
builtin-interfaces==1.6.0
camera-calibration==4.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
composition-interfaces==1.6.0
contourpy==1.3.1
cv-bridge==3.5.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
demo-nodes-py==0.27.2
diagnostic-msgs==5.0.1
domain-coordinator==0.11.2
example-interfaces==0.10.2
examples-rclpy-executors==0.18.1
examples-rclpy-minimal-action-client==0.18.1
examples-rclpy-minimal-action-server==0.18.1
examples-rclpy-minimal-client==0.18.1
examples-rclpy-minimal-publisher==0.18.1
examples-rclpy-minimal-service==0.18.1
examples-rclpy-minimal-subscriber==0.18.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
geometry-msgs==5.0.1
gps-msgs==2.0.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
image-geometry==3.5.0
interactive-markers==2.4.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
laser-geometry==2.5.0
launch==2.0.4
launch-ros==0.24.1
launch-testing==2.0.4
launch-testing-ros==0.24.1
launch-xml==2.0.4
launch-yaml==2.0.4
lifecycle-msgs==1.6.0
logging-demo==0.27.2
map-msgs==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
message-filters==4.7.1
mistune==3.1.1
msgpack==1.0.8
multidict==6.1.0
nav-msgs==5.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
odrive==0.6.9.post0
osrf-pycommon==2.1.2
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.27.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyelftools==0.31
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyserial==3.5
python-can==4.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-qt-binding==1.2.4
pyusb==1.3.1
PyYAML==6.0.2
pyzmq==26.2.1
qt-dotgraph==2.4.3
qt-gui==2.4.3
qt-gui-cpp==2.4.3
qt-gui-py-common==2.4.3
quality-of-service-demo-py==0.27.2
rcl-interfaces==1.6.0
rclpy==4.1.6
rcutils==6.2.3
referencing==0.36.2
requests==2.32.3
resource-retriever==3.2.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rmw-dds-common==2.0.1
ros-gz-interfaces==0.254.2
ros-ign-interfaces==0.254.2
ros2action==0.25.7
ros2bag==0.22.7
ros2bag-mcap-cli==0.22.7
ros2bag-sqlite3-cli==0.22.7
ros2cli==0.25.7
ros2component==0.25.7
ros2doctor==0.25.7
ros2interface==0.25.7
ros2launch==0.24.1
ros2lifecycle==0.25.7
ros2multicast==0.25.7
ros2node==0.25.7
ros2param==0.25.7
ros2pkg==0.25.7
ros2run==0.25.7
ros2service==0.25.7
ros2topic==0.25.7
rosbag2-interfaces==0.22.7
rosbag2-py==0.22.7
rosgraph-msgs==1.6.0
rosidl-adapter==4.0.1
rosidl-cli==4.0.1
rosidl-cmake==4.0.1
rosidl-generator-c==4.0.1
rosidl-generator-cpp==4.0.1
rosidl-generator-py==0.18.0
rosidl-generator-type-description==4.0.1
rosidl-parser==4.0.1
rosidl-pycommon==4.0.1
rosidl-runtime-py==0.12.0
rosidl-typesupport-c==3.0.1
rosidl-typesupport-cpp==3.0.1
rosidl-typesupport-fastrtps-c==3.0.2
rosidl-typesupport-fastrtps-cpp==3.0.2
rosidl-typesupport-introspection-c==4.0.1
rosidl-typesupport-introspection-cpp==4.0.1
rpds-py==0.22.3
rpyutils==0.3.2
rqt-action==2.1.2
rqt-bag==1.3.5
rqt-bag-plugins==1.3.5
rqt-console==2.1.1
rqt-graph==1.4.3
rqt-gui==1.3.4
rqt-gui-py==1.3.4
rqt-msg==1.3.1
rqt-plot==1.2.3
rqt-publisher==1.6.3
rqt-py-common==1.3.4
rqt-py-console==1.1.1
rqt-reconfigure==1.3.4
rqt-service-caller==1.1.1
rqt-shell==1.1.1
rqt-srv==1.1.1
rqt-topic==1.6.1
Send2Trash==1.8.3
sensor-msgs==5.0.1
sensor-msgs-py==5.0.1
service-msgs==1.6.0
shape-msgs==5.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sros2==0.11.3
stack-data==0.6.3
statistics-msgs==1.6.0
std-msgs==5.0.1
std-srvs==5.0.1
stereo-msgs==5.0.1
teleop-twist-keyboard==2.4.0
terminado==0.18.1
tf2-geometry-msgs==0.31.8
tf2-kdl==0.31.8
tf2-msgs==0.31.8
tf2-py==0.31.8
tf2-ros-py==0.31.8
tf2-sensor-msgs==0.31.8
tf2-tools==0.31.8
theora-image-transport==3.1.0
tinycss2==1.4.0
tomli==2.2.1
topic-monitor==0.27.2
tornado==6.4.2
traitlets==5.14.3
trajectory-msgs==5.0.1
turtlesim==1.6.1
type-description-interfaces==1.6.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
unique-identifier-msgs==2.3.2
uri-template==1.3.0
urllib3==2.3.0
vision-msgs==4.1.1
visualization-msgs==5.0.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xacro==2.0.9
yarl==1.18.3
